# Defer `fractions.Fraction` import to save 5ms